adds contact page
contact page has a form to send email;
adds support to send email;
changes routes to company, we were getting just the first company,
so we made a route to make it easy but when we were trying to
edit, the URL was '/company.1' not '/company/1' and it was causing
some trouble;
creates an index for company, it won't be used, but just in case
someone access it from the url. Maybe could be restricted in routes.